+++ app/javascript/gobierto_investments/webapp/containers/home/Aside.vue
@@ -5,34 +5,20 @@
         :key="filter.key"
         class="investments-home-aside--block"
       >
-        <BlockFilter
-          :filter="filter"
-          @set-filter="setFilter"
-        />
+        <slot :filter="filter" />
       </div>
     </template>
   </aside>
 </template>
 
 <script>
-import BlockFilter from "../../components/BlockFilter.vue";
-
 export default {
   name: "HomeAside",
-  components: {
-    BlockFilter
-  },
   props: {
     filters: {
       type: Array,
       default: () => []
     }
-  },
-  methods: {
-    setFilter(filter, key, values) {
-      this.$emit('set-filter', filter, key, values)
-    }
   }
 };
 </script>
-
